Rename `*_trigger` modules into `*_event` modules
No type or variable name changes.
