+++ tests/client_event.rs
